initial commit
This classification has been conducted on such a dataset. All images have been taken in a particular size. Then to reduce dimensions of the picture model uses the principle component analysis technique. It is necessary to use some preprocess techniques as there are many image preprocessing techniques like image restoration, image editing, wavelet transform, etc, therefore it becomes a little bit difficult to select a particular technique. The model is prepared with the experimental observation at the hidden layers so as to adjust the accuracy and to predict well.